+++ src/moops/__init__.py
@@ -1,6 +1,6 @@
 from ._run import run
 from .group import Group
-from .interface import Interface
+from .interface import FileBrowserWithInitialSelection, Interface
 from .presets import Presets
 
-__all__ = ["Group", "Interface", "Presets", "run"]
+__all__ = ["FileBrowserWithInitialSelection", "Group", "Interface", "Presets", "run"]